+++ packages/react/tsconfig.json
@@ -11,6 +11,5 @@
     "src/**/*.ts",
     "src/**/*.tsx",
     "script/**/*.ts",
-    "mocks/**/*.tsx",
   ]
 }
